refactor : put custom suggestion in helper
